Add wait time to loopback test to fix a bug
If you enable recording on a slot soon after changing the direction
of the module connected to that slot, you might get some lingering
(invalid) ADC samples written to the FIFO.  This won't happen in
real life because the module configuration is static after powerup.
It's only a testbench issue.