Clean up constraints
This commit explicitly defines the various clock domains (audio, SPI, etc.) and
constrains the inputs/outputs as a reasonable first pass.  I wrote these up
in response to the various warnings and timing violations flagged by Vivado.
Hopefully a reasonable first pass to resume hardware testing.

A couple of clock nets had to be marked "keep" in order for the constraints
to be applied properly.  This prevents optimizations but isn't a big problem
since everything is so slow.