Put in safeguards against 0-length memory accesses
This was holding things up in testing, and was easy enough to reproduce
in simulation.  This would happen when the FIFO of samples leading back
to the top level (from an ADC slot) was full; the FIFO arbiter would select
that slot since it knew more samples were available, but then compute
(correctly) a maximum transfer length of 0 because there was nowhere
for the samples to go on the other side.

The fix is easy enough: Just cancel the requests and wait until you
have space freed up to write the samples out (i.e. to the host.)