Fix HWCON serializer timing
There was a potential for hold violations in the per-slot HWCON serializers.
The reason was that the serializer requires the parallel (slow) clock to be
stable at the falling edge of the serial (fast) clock.  But in this context,
the parallel and serial clocks (srclk2_sync and srclk_sync) were both really
registers, and they were both triggered by the same clock edge (sclk).

The fix is to make srclk2_sync change on the negative edge of sclk so it
comes "before" srclk_sync.