Stop FIFO arbiter from going into read mode when FIFO is blocked
Reads have a higher priority than writes, so we would repeatedly
enter the "read" state and then go back to idle.  This would normally
be fine, except that if the host is sending data for a DAC, it will
stall until another sample is consumed on the read side.  Kind of
defeats the purpose of having a big FIFO.