Adjustable sample rate: Add HDL support in slot_controller
We already have a provision for selecting between the 22.5792 MHz
and 24.576 MHz clocks, but the ratio of the MCLK to sample rate
was hardcoded at 256.  It can now be programmed to 128/256/384/512.
This should be sufficient to cover the most common sample rates
of 44.1/88.2/176.4 and 48/96/192 kHz.  The I2S frame length
remains 64 bits in all cases.